Move Highest Capacity local kinds to memkind_local
