+++ include/memkind/internal/memkind_capacity.h
@@ -10,8 +10,6 @@ extern "C" {
 
 extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_OPS;
 extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_PREFERRED_OPS;
-extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_LOCAL_OPS;
-extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_LOCAL_PREFERRED_OPS;
 
 #ifdef __cplusplus
 }
