+++ include/memkind/internal/memkind_local.h
@@ -6,6 +6,11 @@
 extern "C" {
 #endif
 
+#include <memkind.h>
+
+extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_LOCAL_OPS;
+extern struct memkind_ops MEMKIND_HIGHEST_CAPACITY_LOCAL_PREFERRED_OPS;
+
 #ifdef __cplusplus
 }
 #endif
